commands/token: move token commands to separate files
