+++ docs/mmctl.rst
@@ -39,6 +39,7 @@ SEE ALSO
 * `mmctl post <mmctl_post.rst>`_ 	 - Management of posts
 * `mmctl system <mmctl_system.rst>`_ 	 - System management
 * `mmctl team <mmctl_team.rst>`_ 	 - Management of teams
+* `mmctl token <mmctl_token.rst>`_ 	 - manage users' access tokens
 * `mmctl user <mmctl_user.rst>`_ 	 - Management of users
 * `mmctl version <mmctl_version.rst>`_ 	 - Prints the version of mmctl.
 * `mmctl websocket <mmctl_websocket.rst>`_ 	 - Display websocket in a human-readable format
